First step to defining variables in YAML
plot_state in the Analysis Element class uses variables.yml to determine the
name of the variable in each dataset as well as to set the contour properties
of the plots. The collection classes do not take advantage of the names_in_file
list (and I'm not sure it makes sense to do so).

Also, I added dirin to the WOA open_dataset arguments.