Move var_dict from yaml -> collection class
The mapping from generic names (e.g. nitrate) to model-specific names (e.g.
NO3) does not need to be encapsulated in YAML. For now I've hard-coded the
definition for WOA to match what Matt named variables in the data sets that use
the POP_gx1v7 grid... but I'll figure out a way to make it grid-dependent when
we need that ability.

This means that the list of variables to be plotted is controlled at the
analysis element level -- individual collections can not pick and choose what
variables to include.

Note that cached data now writes the var_dict as a json file.